requirements.txt
# Main libraries

Django
Pillow

# Additional libraries

django-unfold